requirements.txt
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cssselect2==0.7.0
defusedxml==0.7.1
idna==3.6
pillow==10.2.0
pycparser==2.21
requests==2.31.0
tinycss2==1.2.1
urllib3==2.1.0
webencodings==0.5.1
